backend/requirements.txt
# Flask核心
Flask==3.0.0
Flask-CORS==4.0.0

# AI服务
google-generativeai==0.3.2
openai==1.6.1

# 图片处理
Pillow==10.1.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0

# 异步支持
gevent==23.9.1

# 数据处理
python-json-logger==2.0.7